moved everything player related in audioplayer package, removed listen package
