+++ app/src/main/java/app/books/tanga/feature/audioplayer/infrastructure/PlayerActions.kt
@@ -1,4 +1,4 @@
-package app.books.tanga.feature.audioplayer
+package app.books.tanga.feature.audioplayer.infrastructure
 
 /**
  * Interface that defines actions that can be performed on the player.
