+++ app/src/main/java/app/books/tanga/navigation/MainNavigationGraph.kt
@@ -3,10 +3,10 @@ package app.books.tanga.navigation
 import androidx.compose.runtime.Composable
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
+import app.books.tanga.feature.audioplayer.fullplayer.playSummaryAudio
+import app.books.tanga.feature.audioplayer.fullplayer.toPlaySummaryAudio
 import app.books.tanga.feature.deleteaccount.deleteAccount
 import app.books.tanga.feature.deleteaccount.toDeleteAccount
-import app.books.tanga.feature.listen.playSummaryAudio
-import app.books.tanga.feature.listen.toPlaySummaryAudio
 import app.books.tanga.feature.profile.privacyAndTerms
 import app.books.tanga.feature.read.readSummaryScreen
 import app.books.tanga.feature.read.toReadSummaryScreen
