+++ app/src/test/java/app/books/tanga/feature/audioplayer/PlayerControllerTest.kt
@@ -2,6 +2,10 @@ package app.books.tanga.feature.audioplayer
 
 import androidx.media3.common.Player
 import androidx.media3.session.MediaController
+import app.books.tanga.feature.audioplayer.infrastructure.PlayerAvailability
+import app.books.tanga.feature.audioplayer.infrastructure.PlayerControllerImpl
+import app.books.tanga.feature.audioplayer.infrastructure.PlayerState
+import app.books.tanga.feature.audioplayer.infrastructure.toMediaItem
 import app.books.tanga.fixtures.Fixtures
 import app.books.tanga.rule.MainCoroutineDispatcherExtension
 import app.cash.turbine.test
